server/requirements.txt
flask
protobuf
cassandra-driver
blist
